Another major ammo update
WHEW, okay
Ammo boxes now display a visual indicator of the ammo stored inside of it.
Ammo is also capable of being broken into two pieces to fit in a smaller box.
Racked munitions support (requires 2 variables, explained below)
Its colorcoded, so
- Blue: normal, unmodified rounds
- Orange: ammunition that isn't a shell, or is broken into two
- Green: boxed ammunition
- Red: special bulk ammo display

The bulk ammo display was made to help with performance, and has a new console command that is archived, it is "ACF_MaxRoundsDisplay"
The entire visual display function is also determined by the "acf_show_entity_info" command

As for the previously mentioned racked munitions support, the 2 new variables are
- ActualLength: the actual length of the munition, in cm
- ActualWidth: the actual width of the munition, in cm
Both of these need to be stored in the munition's round table
This makes it so the munition is stored based on the entire model's size