ACF_GetHitAngle
A speculative fix for ACF_GetHitAngle, the core of the issue can't really be fully fixed due to decimal accuracy errors, however we can check when it errors. Also moves the relevant functions to the namespace, and gets rid of clamps. For some reason the old function seemed to randomly give values way outside of its range, this new function now actually checks for the ricochet angle and then compares angles between the two, outputting half of the angle found for compatibility sake.