E2/SF ACF Class function
Adds a function for both E2 and Starfall that allows one to directly grab the class of an ACF entity, for example using this on a TOW missile ammo crate would return ATGM, as opposed to someone needing to explode the name to collect relevant info.
This can further aid in getting specifics about an entity without having to dissect the name, which was ridiculous to leave as-is.

Usage: Entity:acfClass()
Returns: string Entity.Class (as written by ACF)