Somewhat buff fuel consumption rate
The 27.8 fuel consumption rate is based around a single engine, but a significant number of players use 2 or 3 engines, which are much more difficult to use with this previous fuel rate. Decreasing the fuel rate to some degree helps account for this very common setup while still maintaining the intended effect on fuel balancing.