WIP Workshop update checker
This (should) add a static file to the Workshop version only that contains the current commit hash before the upload process, which is then used to check for updates like normal. Haven't been able to test this from start to finish yet, but we'll see how it goes once everything is ready for release.