Expanded exploitable function detouring
A number of functions have been detoured when used on entities with IsACFEntity set to true on it. Most of these will also call ACF.CheckLegal at the time of use, and will block function use
- SetMass
- SetModel
- SetNoDraw
- SetNotSolid - Still allowed (fading door)
- PhysicsInitSphere
- SetCollisionBounds
- SetCollisionGroup - Still allows entities to be made not collideable (context menu)

Makespherical has also been blocked from use on ACF entities, and ProperClipping will be blocked on ACF entities except for procedural armor

Two new convenience functions have been added for development use, ACF.Contraption.SetMass and ACF.Contraption.SetModel; these will set the appropriate variable in the entity's ACF table, and then set the mass normally. As long as Entity.ACF.Mass/Model has been set on an entity PRIOR to mass/model change, the original functions will still call the above functions
Also got rid of LegalMass being set on most entities as it is not used

There is also a new name-and-shame message which is disabled by default on both server and client. When enabled on the server, any time an entity is marked disabled by a legal check, a message is sent out to all players which, depending on their setting, will have a message in the console saying who, what entity, and what happened to make it disabled.