Add Lua Seat Integration
Add Lua Seat Integration

New crew seating system works like this:

On spawn, a baseplate generates a seat that is unkillable and invisible and you can sit in it to gain immunity from damage. Once you leave the seat you lose this immunity.

On spawn, a crew entity generates a seat that is unkillable but the player still receives damage. This is just to deal with the possibility of multicrew.

Seats can be entered by using Shift + E on any prop on the tank

Seat Alias related functions were either commented out or early returned.