Removed Herobrine
Since we're going a different direction with regards to player survivability with regards to scalable baseplates and crew, there is no need to have these alias entities anymore. Did keep one of the small improvements from them, which is to filter out the player's own seat if for example they are using a SWEP and shooting from it.

Removed alias entities.