Fix settings not saving in sp/p2p (might be breaking)
These settings need to be stored in two different locations so that they don't overwrite each other when the same machine is both the server and the client. This will probably break existing server settings once as-is, though; we could set the server file to still just be called "stored.json" to avoid this, but it would be clearer to give it an "_sv" suffix in the long run.