Convert In/Out into LocalPlane classes
There may be a way to determine the direction automatically, but I'm not sure... GetAttachment's return type (AngPos) has an angle associated, but I haven't checked if the models actually have proper attachment angles or not.

If they do, this could be consolidated down further (rather than the "determine it off the ID" stuff, which I don't really like as a solution to this). I just wasn't aware of that until after I had already made this commit (and am now amending it to provide further context, and remind myself later to do that)