Localize many more menus
This commit adds localization files to all of the menus that have not been localized before this point: Guns/Ammo, Engines/Fuel, Piledrivers, Procedural Armor, Crew, all four Turrets menus, and the Cutting Torch (because why not). Also redesigned the layout of the Crew menu a bit and added some globals for min/max ductility and a couple of unit conversions.