Beltfed restrictions + RAC AP/HE
For both optimization and balance reasons, the maximum number of guns for belt fed classes has been changed to 4, with the exception of RACs being 2.

And in line with earlier suggestions, we're trialing HE on RACs. A 2 RAC limit is roughly equivalent to 8 LACs, so this is still a net improvement in reducing the number of HE rounds in the air. Furthermore this option is more optimized than having a tone of guns.