Can we get much higher
Engine RPM will not go below IdleRPM. Very large water jets that require alot of torque will still  be spun at said IdleRPM, which gives you free performance just for a larger propeller.

Until this is fixed in the future it is what it is. I am limitting the size to 1 for now.