Revert RAC ROF and remove HE
RACs previously had their limits reduced from 16 to 2 due to server performance and balance concerns, alongside other automatic weapons.

HE was added to RACs due to request and to see whether it could be properly balanced given the new limits.

It has become apparent that HE on RACs is the worst offender of what is a problematic HE model.

There have also been complaints about RACs losing their niche of high rate of fire, which is a valid point.

Reverting the RAC ROF to 2000 and removing HE is an acceptable compromise to deal with both of these issues.