Revert barrel blocking check for now
In order to improve QOL (guns near each other shooting each other off, guns with muzzles close to the armor colliding with the armor, etc.), we filtered out the entire contraption.

To address the issue of people being able to shoot through their armor, a check was added to determine if the muzzle was blocked. If so, it would blow the gun up.

The check has some issues being consistent, so we are removing it for now.

It is still a valid concern that people may shoot through their armor, so this issue should be revisitted in the future.